[ivp] Apply PVS-Studio fixes
- Fix undefined behavior in ivu_vector insert_after (elems[j] = elems[--j])
- Initialize m_client_data in hk_Constraint constructor
- Initialize m_active in hk_Local_Constraint_System_BP constructor
- Initialize m_min_length in hk_Stiff_Spring_BP constructor
- Initialize force_convex_hull in IVP_Template_Surbuild_LedgeSoup constructor
- Initialize m_joint_friction and m_desired_velocity in hk_Constraint_Limit_BP constructor
- Add virtual destructor to IVP_Liquid_Surface_Descriptor
- Initialize sim_unit in IVP_Event_Sim constructors

Ported from Source-Authors/source-physics commit 8ca257b.

+++ ivp_collision/ivp_mindist_minimize.hxx
@@ -14,8 +14,8 @@ class IVP_Cache_Ledge_Point;
 #if defined(IVP_LOOP_LIST_SIZE)
 struct IVP_MM_Loop_Hash_Struct
 {
-    int a;
-    int b;
+    intp a;
+    intp b;
 };
 #endif
